Move visualizer.py to yamaopt dir from example dir
