+++ example/example.py
@@ -3,7 +3,7 @@ import argparse
 import numpy as np
 from yamaopt.solver import KinematicSolver, SolverConfig
 from yamaopt.polygon_constraint import polygon_to_trans_constraint
-from visualizer import VisManager
+from yamaopt.visualizer import VisManager
 
 if __name__=='__main__':
     parser = argparse.ArgumentParser()
