+++ example/example_multi.py
@@ -5,7 +5,7 @@ import numpy as np
 from skrobot.coordinates.math import rotation_matrix
 from yamaopt.solver import KinematicSolver, SolverConfig
 from yamaopt.polygon_constraint import polygon_to_trans_constraint
-from visualizer import VisManager
+from yamaopt.visualizer import VisManager
 
 if __name__=='__main__':
     parser = argparse.ArgumentParser()
